requirements.txt
alembic==1.6.5
appdirs==1.4.4
certifi==2021.5.30
click==8.0.1
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-Assets==2.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.18
virtualenv==20.4.7
webassets==2.0
Werkzeug==2.0.1
